requirements.txt
Django==1.9.4
django-admin-sortable==2.0.15
Pillow==3.1.1
